requirements.txt
python-dotenv
langchain
langchain-openai
langchain-huggingface
numpy
scipy
torch
sentence_transformers
transformers
bitsandbytes
accelerate
torchvision
peft
einops
timm
flash_attn==1.0.9
wheel
sentencepiece
matplotlib
opencv-python
